refac: split up return type and parameter type refactor
